fix(client): remove unnecessary cast in window resize
Removed redundant integer cast when calculating new window height during
resize. The division already results in an integer due to variable
types,
improving code clarity.

+++ client/main.cpp
@@ -10,7 +10,7 @@ int main() {
   while (!renderer.shouldClose()) {
     if (IsWindowResized()) {
       int newWidth = GetScreenWidth();
-      int newHeight = (int)(newWidth / aspectRatio);
+      int newHeight = newWidth / aspectRatio;
       SetWindowSize(newWidth, newHeight);
     }
     renderer.beginDrawing();
